Revert "Develop"

+++ Events.Api/Events.Infrastructure/Repositories/EventRepository/Implementations/EventRepository.cs
@@ -1,4 +1,5 @@
-﻿using Events.Infrastructure.Entities;
+﻿using Events.Domain.Entities;
+using Events.Domain.Exceptions;
 using Microsoft.EntityFrameworkCore;
 
 namespace Events.Infrastructure.Repositories.EventRepository.Implementations;
@@ -12,41 +13,73 @@ internal class EventRepository : IEventRepository
         this.dbContext = dbContext;
     }
 
-    public async Task Add(EventDb ev)
+    public async Task<bool> Add(Event ev)
     {
+        var eventEntity = await dbContext.Events.Where(e => e.Title == ev.Title).FirstOrDefaultAsync();
+
+        if (eventEntity != null)
+        {
+            throw new ItemAlreadyAddedException("Event");
+        }
+
         dbContext.Events.Add(ev);
         await dbContext.SaveChangesAsync();
+        return true;
     }
 
-    public async Task Delete(EventDb eventEntity)
+    public async Task Delete(string eventId)
     {
+        var eventEntity = await dbContext.Events.FindAsync(eventId);
+
+        if (eventEntity == null)
+        {
+            throw new ItemNotFoundException("Event");
+        }
+
         dbContext.Events.Remove(eventEntity);
         await dbContext.SaveChangesAsync();
     }
 
-    public IQueryable<EventDb> GetAll()
+    public IQueryable<Event> GetAll()
     {
         return dbContext.Events;
     }
 
-    public async Task<EventDb?> GetById(string id)
+    public async Task<Event?> GetById(string id)
     {
         return await dbContext.Events.FindAsync(id);
     }
 
-    public async Task<EventDb?> GetByIdWithRegistrations(string id)
+    public async Task<Event?> GetByIdWithRegistrations(string id)
     {
-        return await dbContext.Events.Include(e => e.Registrations).FirstOrDefaultAsync(e => e.Id == id);
+        var eventInstance = await dbContext.Events.Include(e => e.Registrations).FirstOrDefaultAsync(e => e.Id == id);
+        return eventInstance;
+
     }
 
-    public async Task<EventDb?> GetByTitle(string title)
+    public async Task<Event?> GetByName(string name)
     {
-        return await dbContext.Events.FirstOrDefaultAsync(e => e.Title == title);
+        var eventEntity = await dbContext.Events.Where(e => e.Title == name).FirstOrDefaultAsync();
+
+        return eventEntity;
     }
 
-    public async Task Update(EventDb ev)
+    public async Task Update(Event ev)
     {
-        dbContext.Events.Update(ev);
+        var eventEntity = await dbContext.Events.FindAsync(ev.Id);
+
+        if (eventEntity == null)
+        {
+            throw new ItemNotFoundException("Event");
+        }
+
+        eventEntity.Title = ev.Title;
+        eventEntity.Describtion = ev.Describtion;
+        eventEntity.ImageUrl = ev.ImageUrl;
+        eventEntity.CategoryId = ev.CategoryId;
+        eventEntity.Date = ev.Date;
+        eventEntity.Place = ev.Place;
+
         await dbContext.SaveChangesAsync();
     }
 }
